#1823 Move KernelModuleInlineTrans to a single file and idem for its tests
