+++ src/psyclone/tests/domain/gocean/transformations/gocean1p0_transformations_test.py
@@ -53,11 +53,10 @@ from psyclone.psyir.transformations import LoopFuseTrans, LoopTrans, \
     TransformationError
 from psyclone.transformations import ACCKernelsTrans, ACCRoutineTrans, \
     OMPParallelTrans, GOceanOMPParallelLoopTrans, GOceanOMPLoopTrans, \
-    KernelModuleInlineTrans, OMPLoopTrans, ACCParallelTrans, \
-    ACCEnterDataTrans, ACCLoopTrans
+    OMPLoopTrans, ACCParallelTrans, ACCEnterDataTrans, ACCLoopTrans
 from psyclone.domain.gocean.transformations import GOConstLoopBoundsTrans
 from psyclone.tests.gocean_build import GOceanBuild
-from psyclone.tests.utilities import count_lines, get_invoke, Compile
+from psyclone.tests.utilities import count_lines, get_invoke
 
 # The version of the PSyclone API that the tests in this file
 # exercise
@@ -991,126 +990,6 @@ def test_module_noinline_default(tmpdir):
     # inlining)
     assert 'USE compute_cu_mod, ONLY: compute_cu_code' in gen
     assert GOceanBuild(tmpdir).code_compiles(psy)
-
-
-def test_module_inline(tmpdir):
-    ''' Test that we can succesfully inline a basic kernel subroutine
-    routine into the PSy layer module by directly setting inline to
-    true for the specified kernel. '''
-    psy, invoke = get_invoke("single_invoke_three_kernels.f90", API, idx=0,
-                             dist_mem=False)
-    schedule = invoke.schedule
-    kern_call = schedule.children[0].loop_body[0].loop_body[0]
-    kern_call.module_inline = True
-    gen = str(psy.gen)
-    # check that the subroutine has been inlined correctly
-    expected = (
-        "    END SUBROUTINE invoke_0\n"
-        "    SUBROUTINE compute_cu_code(i, j, cu, p, u)\n")
-    assert expected in gen
-    # check that the associated use no longer exists
-    assert 'USE compute_cu_mod, ONLY: compute_cu_code' not in gen
-    assert GOceanBuild(tmpdir).code_compiles(psy)
-
-
-def test_module_inline_with_transformation(tmpdir):
-    ''' Test that we can succesfully inline a basic kernel subroutine
-    routine into the PSy layer module using a transformation '''
-    psy, invoke = get_invoke("single_invoke_three_kernels.f90", API, idx=0,
-                             dist_mem=False)
-    schedule = invoke.schedule
-    kern_call = schedule.children[1].loop_body[0].loop_body[0]
-    inline_trans = KernelModuleInlineTrans()
-    inline_trans.apply(kern_call)
-    gen = str(psy.gen)
-    # check that the subroutine has been inlined
-    assert 'SUBROUTINE compute_cv_code(i, j, cv, p, v)' in gen
-    # check that the associated use no longer exists
-    assert 'USE compute_cv_mod, ONLY: compute_cv_code' not in gen
-    assert GOceanBuild(tmpdir).code_compiles(psy)
-
-
-def test_transformation_inline_error_if_not_kernel():
-    ''' Test that the inline transformation fails if the object being
-    passed is not a kernel'''
-    _, invoke = get_invoke("single_invoke_three_kernels.f90", API, idx=0,
-                           dist_mem=False)
-    schedule = invoke.schedule
-    kern_call = schedule.children[0].loop_body[0]
-    inline_trans = KernelModuleInlineTrans()
-    with pytest.raises(TransformationError):
-        inline_trans.apply(kern_call)
-
-
-def test_module_inline_with_sub_use(tmpdir):
-    ''' Test that we can module inline a kernel subroutine which
-    contains a use statement'''
-    psy, invoke = get_invoke("single_invoke_scalar_int_arg.f90", API, idx=0,
-                             dist_mem=False)
-    schedule = invoke.schedule
-    kern_call = schedule.children[0].loop_body[0].loop_body[0]
-    inline_trans = KernelModuleInlineTrans()
-    inline_trans.apply(kern_call)
-    gen = str(psy.gen)
-    # check that the subroutine has been inlined
-    assert 'SUBROUTINE bc_ssh_code(ji, jj, istep, ssha, tmask)' in gen
-    # check that the use within the subroutine exists
-    assert 'USE grid_mod' in gen
-    # check that the associated psy use does not exist
-    assert 'USE bc_ssh_mod, ONLY: bc_ssh_code' not in gen
-    assert GOceanBuild(tmpdir).code_compiles(psy)
-
-
-def test_module_inline_same_kernel(tmpdir):
-    '''Tests that correct results are obtained when an invoke that uses
-    the same kernel subroutine more than once has that kernel
-    inlined'''
-    psy, invoke = get_invoke("test14_module_inline_same_kernel.f90", API,
-                             idx=0)
-    schedule = invoke.schedule
-    kern_call = schedule.coded_kernels()[0]
-    inline_trans = KernelModuleInlineTrans()
-    inline_trans.apply(kern_call)
-    gen = str(psy.gen)
-    # check that the subroutine has been inlined
-    assert 'SUBROUTINE compute_cu_code(' in gen
-    # check that the associated psy "use" does not exist
-    assert 'USE compute_cu_mod, ONLY: compute_cu_code' not in gen
-    # check that the subroutine has only been inlined once
-    count = count_lines(gen, "SUBROUTINE compute_cu_code(")
-    assert count == 1, "Expecting subroutine to be inlined once"
-    assert GOceanBuild(tmpdir).code_compiles(psy)
-
-
-def test_module_inline_and_compile(tmpdir):
-    '''ATM incorrect code is produced if a kernel is inlined, that
-    uses variable from the original module. Proper solution would
-    likely be to add a 'use' statement to the inline kernel (which
-    again only works if the module variable is accessible outside
-    of the module)
-    '''
-    Compile.skip_if_compilation_disabled()
-    psy, invoke = get_invoke("test14_module_inline_same_kernel.f90", API,
-                             idx=0, dist_mem=False)
-    schedule = invoke.schedule
-    kern_call = schedule.children[0].loop_body[0].loop_body[0]
-    inline_trans = KernelModuleInlineTrans()
-    inline_trans.apply(kern_call)
-    assert GOceanBuild(tmpdir).code_compiles(psy)
-
-
-def test_module_inline_warning_no_change():
-    '''Test of the warning clause in the Kernel transformation when
-    no change is made to the inlining of a Kernel i.e. the inlining
-    request is already what is happening. No warning is currently made
-    as we have not added logging to the code but this test covers the
-    clause '''
-    _, invoke = get_invoke("test14_module_inline_same_kernel.f90", API, idx=0,
-                           dist_mem=False)
-    schedule = invoke.schedule
-    kern_call = schedule.coded_kernels()[0]
-    inline_trans = KernelModuleInlineTrans()
-    inline_trans.apply(kern_call, {"inline": False})
 
 
 def test_acc_parallel_not_a_loop():
