+++ src/psyclone/transformations.py
@@ -50,7 +50,7 @@ from psyclone.configuration import Config
 from psyclone.domain.lfric import KernCallArgList, LFRicConstants
 from psyclone.dynamo0p3 import DynHaloExchangeEnd, DynHaloExchangeStart, \
     DynInvokeSchedule, DynKern
-from psyclone.errors import InternalError, GenerationError
+from psyclone.errors import InternalError
 from psyclone.gocean1p0 import GOInvokeSchedule
 from psyclone.nemo import NemoInvokeSchedule
 from psyclone.psyGen import Transformation, CodedKern, Kern, InvokeSchedule, \
@@ -58,7 +58,7 @@ from psyclone.psyGen import Transformation, CodedKern, Kern, InvokeSchedule, \
 from psyclone.psyir.nodes import ACCDataDirective, ACCDirective, \
     ACCEnterDataDirective, ACCKernelsDirective, ACCLoopDirective, \
     ACCParallelDirective, ACCRoutineDirective, Assignment, CodeBlock, \
-    Directive, KernelSchedule, Loop, Node, OMPDeclareTargetDirective, \
+    Directive, Loop, Node, OMPDeclareTargetDirective, \
     OMPDirective, OMPDoDirective, OMPLoopDirective, OMPMasterDirective, \
     OMPParallelDirective, OMPParallelDoDirective, OMPSerialDirective, \
     OMPSingleDirective, OMPTaskloopDirective, PSyDataNode, Reference, \
@@ -1115,163 +1115,6 @@ class ColourTrans(LoopTrans):
         # pylint: disable=no-self-use
         raise InternalError("_create_colours_loop() must be overridden in an "
                             "API-specific sub-class.")
-
-
-class KernelModuleInlineTrans(Transformation):
-    ''' Module-inlines (bring the subroutine to the same compiler-unit) the
-    subroutine pointed by this Kernel. For example:
-
-    >>> invoke = ...
-    >>> schedule = invoke.schedule
-    >>>
-    >>> inline_trans = KernelModuleInlineTrans()
-    >>>
-    >>> inline_trans.apply(schedule.children[0].loop_body[0])
-    >>> # Uncomment the following line to see a text view of the schedule
-    >>> # print(schedule.view())
-
-    .. warning ::
-        For this transformation to work correctly, the Kernel subroutine
-        must only use data that is passed in by argument, declared locally
-        or included via use association within the subroutine. Two
-        examples where in-lining will not work are:
-
-        #. A variable is declared within the module that ``contains`` the
-           Kernel subroutine and is then accessed within that Kernel;
-        #. A variable is included via use association at the module level
-           and accessed within the Kernel subroutine.
-
-        The transformation will reject attempts to in-line such kernels.
-    '''
-
-    def __str__(self):
-        return ("Inline a kernel subroutine into the PSy module")
-
-    @property
-    def name(self):
-        ''' Returns the name of this transformation as a string.'''
-        return "KernelModuleInline"
-
-    def validate(self, node, options=None):
-        '''
-        Checks that the supplied node is a Kernel and that it is possible to
-        inline the PSyIR of its contents.
-
-        :param kern: the kernel which is the target of the transformation.
-        :type kern: :py:class:`psyclone.psyGen.Kern` or sub-class
-        :param options: a dictionary with options for transformations.
-        :type options: dictionary of string:values or None
-
-        :raises TransformationError: if the target node is not a sub-class of \
-                                     psyGen.Kern.
-        :raises TransformationError: if the subroutine containing the \
-                                     implementation of the kernel cannot be \
-                                     retrieved wiht 'get_kernel_schedule'.
-        :raises TransformationError: if the kernel cannot be safely inlined.
-
-        '''
-
-        if not isinstance(node, Kern):
-            raise TransformationError(
-                f"Target of a kernel transformation must be a sub-class of "
-                f"psyGen.Kern but got '{type(node).__name__}'")
-
-        # Check that the PSyIR and associated Symbol table of the Kernel is OK.
-        # If this kernel contains symbols that are not captured in the PSyIR
-        # SymbolTable then this raises an exception.
-        try:
-            kernel_schedule = node.get_kernel_schedule()
-        except Exception as error:
-            raise TransformationError(
-                f"{self.name} failed to retrieve PSyIR for kernel "
-                f"'{node.name}' using the 'get_kernel_schedule' method."
-                ) from error
-
-        # Check that all kernel symbols are declared in the kernel
-        # symbol table(s). At this point they may be declared in a
-        # container containing this kernel which is not supported.
-        for var in kernel_schedule.walk(Reference):
-            try:
-                var.scope.symbol_table.lookup(
-                    var.name, scope_limit=var.ancestor(KernelSchedule))
-            except KeyError as err:
-                raise TransformationError(
-                    f"Kernel '{node.name}' contains accesses to data (variable"
-                    f" '{var.name}') that are not present in the Symbol Table"
-                    f"(s) within KernelSchedule scope. Cannot inline such a"
-                    f" kernel.") from err
-
-    def apply(self, node, options=None):
-        ''' Bring the kernel subroutine in this Container.
-
-        :param node: the kernel to module-inline.
-        :type node: :py:class:`psyclone.psyGen.CodedKern`
-        :param options: a dictionary with options for transformations.
-        :type options: dictionary of string:values or None
-
-        '''
-        self.validate(node, options)
-
-        if not options:
-            options = {}
-
-        name = node.name
-        try:
-            existing_symbol = node.scope.symbol_table.lookup(name)
-        except KeyError:
-            existing_symbol = None
-
-        if not existing_symbol:
-            # If it doesn't exist already, module-inline the subroutine by:
-            # 1) Registering the subroutine symbol in the Container
-            from psyclone.psyir.symbols import RoutineSymbol, ContainerSymbol
-            from psyclone.psyir.nodes import Literal, Container, ScopingNode
-            node.ancestor(Container).symbol_table.add(RoutineSymbol(name))
-            # 2) Insert the relevant code into the tree.
-            inlined_code = node.get_kernel_schedule()
-            container = inlined_code.ancestor(Container)
-
-            bring_in = set()
-
-            for literal in inlined_code.walk(Literal):
-                if isinstance(literal.datatype.precision, DataSymbol):
-                    name = literal.datatype.precision.name
-                    # FIXME
-
-            for scope in inlined_code.walk(ScopingNode):
-                for symbol in scope.symbol_table.symbols:
-                    if symbol.is_unresolved:
-                        # We don't know where this comes from, we need to bring
-                        # in all top-level imports
-                        for mod in container.symbol_table.containersymbols:
-                            bring_in.add(mod)
-                    elif symbol.is_import:
-                        pass  # Add to bring in if imported outside?
-                    elif symbol.is_local:
-                        # This should be on the validate
-                        # Ok if is a constant, otherwise it should be an
-                        # # error?
-                        pass
-
-            for symbol in bring_in:
-                if symbol.name in inlined_code.symbol_table:
-                    same_symbol = inlined_code.symbol_table.lookup(symbol.name)
-                    if not isinstance(same_symbol, ContainerSymbol):
-                        raise InternalError(
-                            f"Incoherent PSyIR found ({symbol.name} not as "
-                            f"expected)")
-                else:
-                    inlined_code.symbol_table.add(symbol)
-                    # Note that momentarily we have the same symbol in two
-                    # symbol tables, but this is not a problem because below
-                    # we detach and discard the ancestor part of the tree.
-
-            node.root.addchild(inlined_code.detach())
-
-        else:
-            # The routine symbol already exist, we need to make sure it
-            # references the same routine. (would eq work?)
-            pass
 
 
 class Dynamo0p3ColourTrans(ColourTrans):
@@ -3171,7 +3014,6 @@ __all__ = ["OMPLoopTrans",
            "Dynamo0p3OMPLoopTrans",
            "GOceanOMPLoopTrans",
            "ColourTrans",
-           "KernelModuleInlineTrans",
            "Dynamo0p3ColourTrans",
            "ParallelRegionTrans",
            "OMPSingleTrans",
